requirements.txt
selenium==3.141.0
allure-pytest==2.7.0
pytest==5.0.0